Add a "serve" command
This command runs the web server in the backend, but does not attempt to open the browser, making it ideal for local development. Most likely runbook authors and runbook consumers won't make use of this.

Also, refactor so that the backend functionality is all in a dedicated package and an individual comd need only call the relevant backend functions.