Add duplicate component warnings and execution mode support
We now use the executable registry for `runbooks open` and `runbooks serve`, but optionally use "live-file-reload" for `runbooks watch`. This creates a secure experience for consumers and developers and a convenient experience for authors, with an option to add more security in exchange for convenience.

This commit also added a warning banner in case the Executable Registry detects confusing behavior, such as two components with the same ID, which means that one component's scripts will be ignored.