Import only variable values for Check and Command BoilerplateInputsId references.
Previously, Command and Check blocks would attempt to render a complete boilerplate template when it had a non-empty value for BoilerplateInputsId. But that caused problems when it referenced a BoilerplateInputs that had a non-trivial boilerplate template (e.g. including dependencies).

Now the Command or Check block imports the variable values, but ignores the boilerplate.yml contents otherwise.